// src/types/producers.types.ts
export type TipoPersona = "FISICA" | "JURIDICA";
// ...
export interface ProductionCompanyByIdResponse {
  productora: {
    id_productora?: string;
    usuarioPrincipal: {
      apellido: string;
      email: string;
      id_usuario: string;
      nombre: string;
    };
    cuit_cuil: string;
    nombre_productora: string;
    tipo_persona: TipoPersona;
    apellidos: string;
    nombres: string;
    email: string;
    calle: string;
    numero: string;
    ciudad: string;
    localidad: string;
    provincia: string;
    codigo_postal: string;
    telefono: string;
    nacionalidad: string;
    alias_cbu: string;
    cbu: string;
    cantidad_fonogramas: number;
    razon_social: string;
    denominacion_sello: string;
    datos_adicionales: string;
    fecha_alta: Date;
    fecha_ultimo_fonograma: Date;
    apellidos_representante: string;
    nombres_representante: string;
    cuit_representante: string;
    createdAt: Date;
    updatedAt: Date;
    codigosDeLaProductora?: {
      tipo: string;
      codigo_productora: string;
    }[];
  };
}

export interface GetCompaniesResponse {
  total: number;
  totalPages: number;
  currentPage: number;
  data: ProductionCompany[];
}

export interface UpdateProducerPayload {
  nombre_productora: string;
  tipo_persona: TipoPersona;
  telefono: string;
  email: string;
  cuit_cuil: string;
  calle: string;
  numero: string;
  ciudad: string;
  localidad: string;
  provincia: string;
  codigo_postal: string;
  nacionalidad: string;
  datos_adicionales?: string;
  denominacion_sello?: string;
  nombres?: string;
  apellidos?: string;
  razon_social?: string;
  nombres_representante?: string;
  apellidos_representante?: string;
  cuit_representante?: string;
  cbu?: string;
  alias_cbu?: string;
}
// ...
export const PROVINCIAS = [
  "Buenos Aires",
  "CABA",
  "Catamarca",
  "Chaco",
  "Chubut",
  "Córdoba",
  "Corrientes",
  "Entre Ríos",
  "Formosa",
  "Jujuy",
  "La Pampa",
  "La Rioja",
  "Mendoza",
  "Misiones",
  "Neuquén",
  "Río Negro",
  "Salta",
  "San Juan",
  "San Luis",
  "Santa Cruz",
  "Santa Fe",
  "Santiago del Estero",
  "Tierra del Fuego",
  "Tucumán",
] as const;